more accurately count IPs from CIDRs and other expansions
Previously, we simply counted all IPs plus the sizes of all CIDRs
we encountered.  However, it's possible for some of the IPs to
be within a CIDR, or for there to be overlapping CIDRs.

With this commit, we now laboriously keep track of all CIDRs,
add individual IPs as /32 or /128 CIDRs, and then dedupe the
CIDRs before counting the total IPs.

Yes, this is a whole lot of change for something that rarely
happens.