allow the user to specify resolvers by name
This addresses https://github.com/jschauma/spf/issues/4 and allows
'spf -r 2620:fe::fe' as well as 'spf -r dns.quad9.net'.

This of course means that we need to use the local default resolver
to resolve the given name before we then use the IP addresses of
the given resolver.